Added a directory "extensions" containing the jQuery library as well as two variants of an "AutocompleteCellEditor" each based on a different autocomplete jQuery plugin.
The model is updated (and the modelChanged event fired) when a suggestion is selected either by clicking or by pressing ENTER on it.
I had to move the call to displayEditor after the setup of the onblur callback.
Indeed, the default onblur event handling must be disabled in order to get the "click on a suggestion" working.

The demo uses AutocompleteCellEditor to add auto complete on column "FIRSTNAME".
For testing purposes, demo.html uses the first variant, and demo_attach.html the second one.

This is based on an original idea by Jasper Visser.