Client side pagination (BETA, not tested in "attach" mode). Added "pageSize" option: if non-zero, it means we want to use pagination.
Pagination API:
- getNbPages()
- getCurrentPageIndex()
- setPageIndex(pageIndex)

Sorting and filtering applies to the whole model, not just the current page.
We thus re-render the grid after a sort or filter (using new function refreshGrid).
When a filter is applied, we also come back to the first page.

Also, since we now actually filter the model (and not just the view), charts be filtered too (by rendering the chart in the tableFiltered callback).