1/ renderBarCharts can interpret row attributes named "reference_columns"
This should be an object with following properties:
- "column": name or index of column in which to get the reference value for row
- "style": array of SVG style attributes (eg. stroke, stroke-width, stroke-dasharray, zIndex).
  default attribute values are set so "style" can be null

An horizontal reference line is then drawed with given style at Y = reference value.
The width of each line is computed to fill all space between two columns.
Thus the line will be a long continuous line if reference value is the same for all rows.

This is done in the "on complete" callback of Highcharts.

2/ Tested combination of a line/scatter serie in column chart (commented out)