Added config options editNextOnTab (true), editNextOnEnter (false) and editNextDirection (horizontal).
It is thus possible to automatically go to edit the next cell also when pressing ENTER (and not only with TAB) and to go downwards (instead of right).

We ignore columns of type "date"  when looking for the next editable cell.

Removed useless code supposed to stop when coming back to the originally edited cell (we don't know which was the original cell).